+++ algebra/u256/src/conversion.rs
@@ -1,0 +1,164 @@
+use crate::{
+    U256,
+    commutative_binop,
+    division::{divrem_nby1, divrem_nbym},
+    gcd::inv_mod,
+    noncommutative_binop,
+    utils::{adc, div_2_1, mac, sbb},
+};
+use std::{
+    cmp::Ordering,
+    num::Wrapping,
+    ops::{
+        Add, AddAssign, BitAnd, BitAndAssign, Div, DivAssign, Mul, MulAssign, Rem, RemAssign, Shl,
+        ShlAssign, Shr, ShrAssign, Sub, SubAssign,
+    },
+    prelude::v1::*,
+    u64,
+};
+
+impl U256 {
+    pub fn from_bytes_be(n: &[u8; 32]) -> Self {
+        Self::from_limbs(
+            u64::from_be_bytes([n[24], n[25], n[26], n[27], n[28], n[29], n[30], n[31]]),
+            u64::from_be_bytes([n[16], n[17], n[18], n[19], n[20], n[21], n[22], n[23]]),
+            u64::from_be_bytes([n[8], n[9], n[10], n[11], n[12], n[13], n[14], n[15]]),
+            u64::from_be_bytes([n[0], n[1], n[2], n[3], n[4], n[5], n[6], n[7]]),
+        )
+    }
+
+    pub fn to_bytes_be(&self) -> [u8; 32] {
+        let mut r = [0; 32];
+        let mut n = self.clone();
+        // We want truncation here
+        #[allow(clippy::cast_possible_truncation)]
+        for i in (0..32).rev() {
+            r[i] = n.limb(0) as u8;
+            n >>= 8;
+        }
+        r
+    }
+}
+
+macro_rules! impl_from_uint {
+    ($type:ty) => {
+        impl From<$type> for U256 {
+            // $type could be u64, which triggers the lint.
+            #[allow(trivial_numeric_casts)]
+            fn from(n: $type) -> Self {
+                Self::from_limbs(n as u64, 0, 0, 0)
+            }
+        }
+    };
+}
+
+impl_from_uint!(u8);
+impl_from_uint!(u16);
+impl_from_uint!(u32);
+impl_from_uint!(u64);
+impl_from_uint!(usize);
+
+impl From<u128> for U256 {
+    fn from(n: u128) -> Self {
+        // We want truncation here
+        #[allow(clippy::cast_possible_truncation)]
+        Self::from_limbs(n as u64, (n >> 64) as u64, 0, 0)
+    }
+}
+
+macro_rules! impl_from_int {
+    ($t:ty) => {
+        impl From<$t> for U256 {
+            // We want twos-complement casting
+            #[allow(clippy::cast_sign_loss)]
+            // We want truncation here
+            #[allow(clippy::cast_possible_truncation)]
+            fn from(n: $t) -> Self {
+                if n >= 0 {
+                    Self::from_limbs(n as u64, 0, 0, 0)
+                } else {
+                    Self::from_limbs(
+                        n as u64,
+                        u64::max_value(),
+                        u64::max_value(),
+                        u64::max_value(),
+                    )
+                }
+            }
+        }
+    };
+}
+
+impl_from_int!(i8);
+impl_from_int!(i16);
+impl_from_int!(i32);
+impl_from_int!(i64);
+impl_from_int!(isize);
+
+impl From<i128> for U256 {
+    // We want twos-complement casting
+    #[allow(clippy::cast_sign_loss)]
+    // We want truncation here
+    #[allow(clippy::cast_possible_truncation)]
+    fn from(n: i128) -> Self {
+        if n >= 0 {
+            Self::from_limbs(n as u64, (n >> 64) as u64, 0, 0)
+        } else {
+            Self::from_limbs(
+                n as u64,
+                (n >> 64) as u64,
+                u64::max_value(),
+                u64::max_value(),
+            )
+        }
+    }
+}
+
+macro_rules! as_int {
+    ($name:ident, $type:ty) => {
+        // $type could be u64, which triggers the lint.
+        #[allow(trivial_numeric_casts)]
+        pub fn $name(&self) -> $type {
+            self.limb(0) as $type
+        }
+    };
+}
+
+// We don't want newlines between the macro invocations.
+#[rustfmt::skip]
+impl U256 {
+    as_int!(as_u8, u8);
+    as_int!(as_u16, u16);
+    as_int!(as_u32, u32);
+    as_int!(as_u64, u64);
+    as_int!(as_usize, usize);
+    as_int!(as_i8, i8);
+    as_int!(as_i16, i16);
+    as_int!(as_i32, i32);
+    as_int!(as_i64, i64);
+    as_int!(as_isize, isize);
+
+    // Clippy is afraid that casting u64 to u128 is lossy
+    #[allow(clippy::cast_lossless)]
+    pub fn as_u128(&self) -> u128 {
+        (self.limb(0) as u128) | ((self.limb(1) as u128) << 64)
+    }
+
+    // Clippy is afraid that casting u64 to u128 is lossy
+    #[allow(clippy::cast_lossless)]
+    pub fn as_i128(&self) -> i128 {
+        (self.limb(0) as i128) | ((self.limb(1) as i128) << 64)
+    }
+}
+
+// TODO: Replace literals with u256h!
+#[allow(clippy::unreadable_literal)]
+// Quickcheck requires pass by value
+#[allow(clippy::needless_pass_by_value)]
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use quickcheck_macros::quickcheck;
+    use zkp_macros_decl::u256h;
+
+}
